Revert "Second dist subtree commit"
