+++ package.json
@@ -1,0 +1,30 @@
+{
+  "name": "inclement-emerald-pokedex",
+  "version": "1.0.0",
+  "description": "Pokedex application for the Inclement Emerald ROM Hack",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "watch": "webpack --watch",
+    "start": "webpack serve --open"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ydarissep/inclement-emerald-pokedex.git"
+  },
+  "author": "Ydarissep",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/ydarissep/inclement-emerald-pokedex/issues"
+  },
+  "homepage": "https://github.com/ydarissep/inclement-emerald-pokedex#readme",
+  "devDependencies": {
+    "css-loader": "^6.7.1",
+    "html-webpack-plugin": "^5.5.0",
+    "standard": "^17.0.0",
+    "style-loader": "^3.3.1",
+    "webpack": "^5.72.1",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.9.0"
+  }
+}
