+++ src/api/getForms.js
@@ -1,0 +1,8 @@
+async function getForms () {
+  const rawForms = await fetch('https://raw.githubusercontent.com/BuffelSaft/pokeemerald/master/src/data/pokemon/form_species_tables.h')
+  const textForms = await rawForms.text()
+
+  return textForms
+}
+
+export default getForms
