+++ src/api/getTMHMLearnsets.js
@@ -1,0 +1,8 @@
+async function getTMHMLearnsets () {
+  const rawTMHMLearnsets = await fetch('https://raw.githubusercontent.com/BuffelSaft/pokeemerald/master/src/data/pokemon/tmhm_learnsets.h')
+  const textTMHMLearnsets = await rawTMHMLearnsets.text()
+
+  return textTMHMLearnsets
+}
+
+export default getTMHMLearnsets
