+++ src/index.js
@@ -1,0 +1,4 @@
+import './styles/style.css'
+import buildPokemonObj from './helpers/buildPokemonObj'
+
+buildPokemonObj()
